small fixes
- Don't debug print Timer events in src/bin.rs event loop; they're too
frequent and pollute the logs
- chain set_{fg,bg,..} method calls for &mut Cell
- remove unneeded u8 to u8 cast

+++ ui/src/terminal/cells.rs
@@ -740,7 +740,7 @@ impl Color {
             | b @ Color::Cyan
             | b @ Color::White
             | b @ Color::Default => AnsiValue(b.as_byte()),
-            Color::Byte(b) => AnsiValue(b as u8),
+            Color::Byte(b) => AnsiValue(b),
             Color::Rgb(_, _, _) => AnsiValue(0),
         }
     }
